Make PARSE call SUBPARSE "pseudo-native"
This reorganizes parse in such a way that it invokes a native-formatted
routine, so each parse recursion can be given a stack frame.

It also changes the error delivery so that error messages will indicate
their "near" points as originating from the parse instruction that
caused the problem, as opposed to the overall PARSE call.

In the process, this rewrite exposes bugs in handling of modifications
to the parse rules in the process of a parse...which will be addressed
in a separate commit.