Move "safe fetch" to after GROUP! testing
Because all operations processing instructions in PARSE needed to
eventually fetch the next instruction, the code was biased to save the
current rule into `rule` and then fetch up front.  This included cases
of GROUP! or variables that resolved to GROUP!, which may execute code
that would disrupt the state of the fetch.

As a consequence, all fetching had to be deferred until after any
arbitrary user code ran.  This is less convenient, but required.  It
would be possible to go through and group things to use the fetch less
often, or to make a fetch-and-continue, but this gets the tests passing
for now.