Simplify bootstrap, eliminate NATIVE native
A modification to build a full REBVAL table for all the natives that
could be used by the source demonstrated there was an existing flaw
in the usage of loaded natives (as in RETURN, LEAVE, and EVAL) in the
evaluator itself.  These were being consulted before being initialized.

It would be possible to hand-build the functions in the same way that
NATIVE and ACTION were being built.  However, this list could grow, and
it seemed the easiest thing to do would be to just get rid of the idea
that there is a "NATIVE native".  Considering that users couldn't call
the function anyway, it suffices to have it as a notation of expressing
the native definitions instead of needing to be a function itself.

Besides fixing the loophole, this also provides a table so that each
time a new native is needed to be referred to for some reason by value
it does not require adding code into the native registration routine
and the root table.

It also brings online the pending feature of equivalent usermode bodies
for native functions (as a form of documentation).