Macros => inline static functions
This is basically a top-to bottom vetting of the macros, switching
the ones that used an argument more than once over to using an inline
function.  Because `inline static` is used on these definitions, it
means that `inline` could be defined to nothing and they would be
compiled as if a separate copy were statically included in each
header that used it--hence retaining C89 compatibility.

The previous macros did not require the types they worked with to be
defined at the point the macro was defined.  It was only when a macro
was used at a callsite that the definitions would need to be available.
The inline functions however needed much more to be defined.  As a
consequence, this required splitting out the REBVAL and REBSER and
`struct Reb_Frame` definitions into separate headers included earlier
in the ordering of %sys-core.h.  Once all the structures were defined,
the inline functions could be made in a certain ordering.

Since basically every line of %sys-value.h and %sys-series.h needed to
be touched in this change, the opportunity was taken to do some amount
of cleanup in the process.  However, %sys-core.h was reordered in
something of a trial-and-error process to get this to compile.  With
the reorganization strategy now started and working, it now needs
another going-over and review.

squish me