Reduce number of LIB/ROUTINE macros
There were a large number of variations of macros for routines with
VAL_XXX and ROUTINE_XXX and RIN_XXX versions of the same thing.  These
would take a REBVAL*, the other a REBFUN*, and a REBRIN* respectively.

Given the large number of properties to be extracted, it simply makes
more sense for the caller to extract the REBRIN once from a REBVAL*
or a REBFUN* and then make the RIN_XXX calls on that.  This makes that
simplifying step, along with standardizing some of the conventions, e.g.

    CLEAR_RIN_FLAG(r, ROUTINE_FLAG_XXX)

instead of:

    ROUTINE_CLR_FLAG(r, ROUTINE_XXX)

A few similar simplifications to LIB are included.