Streamline routine calls, no frame args copying
Changes to the model in which Rebol acquired parameters and passed them
to the various function dispatchers led to creating an extra copy of
the arguments to call routine.  It turns out that the routine code
wanted a 0-based array of values, while the frame has a 1-based accessor
to the same values.

Switching to use FRM_ARG(f, n) to get the arguments made the copying
unnecessary.

Noticed while tidying the code up, and also using the initialize-at-time
of declaration pattern.  Along with `inline static` this is one of the
compromises agreed upon vs. pure C89 compatibility.