%t-routine.c indentation and spacing adjustment
Ren-C has standardized some formatting conventions, including things
like that cases in switch statements should only not have spacing
between them if they are fallthrough cases.  (This helps notice errors,
and there was a missing `break` in the code actually.)

Also, `case` does not get an additional level of indentation from the
switch.  There were instances of case and a block and the code being
indented, which means the code is indented 8 spaces on a single switch.
That quickly exhausts the horizontal space.

Some tidying included, but nothing semantic.