Buffer FFI args into dedicated series
This streamlines a couple of aspects of the FFI calls.  One is to not
reuse the argument value data bits as the location of the built FFI-args
but to establish a single dedicated `store` series, into which the
appropriately aligned runs of data for each FFI arg are built.

(As a result, variadic parameters can be resident on the data stack
and be converted to their ffi representations without needing pointer
stability for after the conversion...hence they can be dropped without
creating an array to hold them during the call.)

It also makes it clearer what's happening with the return value--in that
it really should only have space created for it to write to, and not
be copied.  Hence `arg_to_ffi` is able to operate on an FFI type alone,
which removes the need to "prep" a phony return argument.