Separate return from args in FFI routines
The convention for ROUTINE! was that there was a single array that
contained both parameters and the return type, with the return type at
index 0.

Beyond being somewhat confusing to read the name "args" and have that
refer not only to inputs but also to an output, there is the additional
problem that every function has a return but not all have arguments.
Also that "void" is permissible only for the return value.  Because the
return had to be treated specially in most all cases, it was confusing
for some parts of the code to be 0-based (not including the return)
with others being 1-based.

This allowed simplifications that mean the fixed argument portion of
the Do_Routine function is shared between variadics and fixed argument
functions.