Drop variadic portion of arg_fftypes after CIF made
Rather than create a full new FFI type argument series on each variadic
call, the routine code would extend the fixed size array.  This saves
on memory and copying while being a little bit sketchy in changing
the routine info structure for a "temporary" effect.

This tightens it up a little bit by asserting the argument series is
in its original condition on each call and restoring it before returning
from the routine call...as opposed to waiting to the next call to tidy
up the change.