Create STU parallel to CTX, FUNC for struct
This switches to using a "canon value array" (as seen in contexts and
functions) for structs.  This means the struct is fully identified by
an array pointer, because the [0]th value is the value of the struct
itself.  This takes over the name for REBSTU and homogenizes the
interface for them to look like REBCTX and REBFUN somewhat.

There are a lot of changes, and this is an "in-progress" snapshot at
the point where varargs.r, gtk.r, and qsort.r are working.