Make struct implicitly creates FFI type
The STRUCT! type essentially only exists for interaction with the FFI.
Hence having a function for making `ffi_type` descriptions separate
from the making of the struct itself doesn't make that much sense.
It is especially problematic due to the lifetime and GC issues that
calling the ffi_type making repeatedly on the same struct causes.

This unifies the ffi_type creation into the making of the struct.