Separate "schema" from structure, mass changes
This is a snap point of a work in progress, to try and shuffle several
things around.  The main thrust is to get a struct to have its ffi
type built at MAKE STRUCT! time and GC'd correctly, so that this is
independent from the REBRIN.

The snapshot is at the point that %gtk.r, %qsort.r, and %varargs.r
are running--with varargs now able to handle structure cast parameters.

It moves the FFI-related definitions into a separate header file whose
inclusion order is *after* sys-value.h.  It also eliminates the
STRUCT_TYPE_XXX enum in favor of embracing FFI_TYPE_XXX and working
around it (rather than duplicating it).  This compromised the REBVAL
passing feature for the moment--more to come.