Unify callback return packing with arg packing
This eliminates some repeated code for turning REBVALs into bits that
the FFI can understand.  The same logic is needed for taking the REBVAL
result from a callback and giving it back to a C function as is needed
when passing a REBVAL argument to a C function by way of a routine.