Unify MAKE and construction syntax, MAKE/TO=>native
A_MAKE and A_TO have historically been unusual actions, in that their
first argument isn't necessarily of the type being acted upon--they
may be a datatype.  This created irregularity in the switch() statement
implementing actions, and the presence of the MAKE and TO code in the
action dispatch also had overlap with the MT_xxx "Make Type" functions.

This switches MAKE and TO to be natives that call their respective
functions.  It tries to be as compatible as possible, while pushing the
code toward greater clarity.  Those instances of actual dependence on
the first argument to MAKE are quarantined into the CONSTRUCT function
and the MAKE wrapper in `<r3-legacy>`.

See #2263 for rationale and consequences behind the unification of
MAKE and construction syntax.