Do not guard port/state
No idea why it needs to be guarded, but removing it doesn't seem
to break anything.

The memory issue can be shown by this example:
>> forever [modified? %/tmp/a.txt]

For every call to modified?, it tries to open a new port, and a new
port/state, which is protected from garbage collection, and thus it
causes a memory leak (even though it's still reachable, but never
free'd).