Restores some authorship information.
During the bulk conversion of file headers some authorship information
was removed from file headers on the basis that it would be contained
in a Credits file instead.

This merges the names from the files along with the contents of the
previous NOTICE file into a unified CREDITS.md.  It is the result of
@hostilefork editing the draft proposal into a form that lists each
person once and splits the corporate support into a separate section.