Implement the REPL's I/O as Rebol code
This revives code from an earlier branch that attempted to push the
REPL implementation into Rebol code.

It works around one of the weaknesses of that approach--which was
debugger interaction.  The problem is that the REPL's stack frames,
being implemented in Rebol code, were becoming wedged in the backtrace
and were difficult to filter out.  This uses the trick of making the
REPL actually only do the I/O and binding to create the block to be
evaluated and returning it to the C code.

Some difficult-to-maintain C code is replaced by this, which had to
deal with problems like parsing the input buffer for multi-line
input.  The trick used is to try LOADing the code, and use the errors
to cue what kind of incompleteness is in effect.  Also, it avoids
having to deal with expanding buffers in C by relying on ordinary
appends of lines to Rebol series.