Comment regarding Ctrl-C during Ctrl-C handling
In R3-Alpha, a "Halted" message is printed out during response to a
Ctrl-C.  This printing code is inside a longjmp handler.  But what
happens if that printing code--itself--is interrupted?

That may seem unlikely, but other "analogous" things could be more
likely.  It should be reviewed, but is not near-term enough to make
it a GitHub issue right now.