Fix tuple conversion between REBVAL and RXIARG
Extensions (RXIARG) expect the tuple data is stored in arg->bytes, and
the first byte is the length of the data. However, there is a size
mismatch between REBTUP (10 bytes) and RXIARG.bytes (8 bytes). But we are
not passing anything but a color tuple to the extension, so this mismatch
doesn't cause any problems.