c++17 patch for const reinterpret_casting
The cast() macro has extra checking in the C++ build.  In C++17 the
rules for how costs can be reinterpret_cast without needing a
const_cast have tightened, so this updates for that.

It also removes the VAL_TYPESET_SYM hack which was motivated by the
conversion of context keys from words to typesets, that has long since
been completed.  It should now be enough to just check the argument is
a typeset (will be done with the inline function changes).