Blocks return void for SELECT, PICK, TAKE...
This commit kicks off the change where an "out of band" void is used
to signal an item could not be picked out of a block, instead of NONE!

As a convenience, the NOT VOID? test (a.k.a. ANY-VALUE?, or to be
VALUE? in the future) is available as ? ... and assignments of void
via a SET-WORD! will not error by default.  This means the following
will work, and is more likely a match for the intent:

    if ? value: select [a b c] 'd [
        ;-- do stuff with value
    ]

Legacy switch retains the old behavior of returning NONE!, although
such a switch may end up being difficult to keep working under
adaptations of mezzanine to the new behavior.