Fix off-by-one problem with typespecs in bootstrap
As of the previous commit, if the executable were tried to use to build
it would build...but the executable would fail.  The reason was a
silent failure of an `append select`, which since the void change
will be a no-op instead of a NONE.  This created fewer typespecs
because 0 had no spec, and an off-by-one error in lining up the specs
with the types.

This starts the typespec array at BLANK instead of a none for REB_0.