REB_NONE => REB_BLANK, NONE-OF plus NEITHER?/NOR
This deprecates NONE under its current meaning to make room for its
interpretation as "none of the conditions".  See the Trello note:

https://trello.com/c/vJTaG3w5

For now, The NONE!, NONE, and NONE? words trigger an error outlining
the change and directing people to BLANK and _.  `<r3-legacy>` defines
them as synonyms for BLANK!, BLANK, and BLANK?.  The interim name for
the planned future functionality is implemented as NONE, but renamed
in the mezzanine to NONE-OF.

This also adds a prefix NEITHER? operator and its infix form NOR.

While removing NONE and #[none] from the code, this uses a mixture of
choices of whether to substitute in BLANK or _.  It's not currently
clear what specific preferences will be on that, so it's just a first
cut as to what might be appealing...since each choice has its benefits.
`return blank` tends to look more literate, while seeing a long list
of uninitialized variables in an object definition tends to scan more
easily when there isn't a word there.