Tweaks to allow -> and <-, eliminate ops.r file
This does a small addition just to permit -> and <-, which are part of
a broader permissiveness of "arrowed" things...based on the idea that
tags with "strange terminals or beginnings" would need to use curly
brace strings, e.g.

    >> print <{- a tag -}>
    <- a tag ->

Overall the priority of making the language more symbolically expressive
is higher than edge-case HTML compatibility (as the source will never
be HTML compatible anyway, and the point was to make something better
in the first place...vs. be crippled by it).