"Endable" replacement for arity hack, infix features
This adds the <end> tag for arguments, and upgrades infix functions
to support handling them.  It also includes fixes for specialized
functions to be run as infix--which requires special handling, as the
specialization may have changed what the "first argument" actually is.

The code is somewhat roughed in at the moment, and asserts in places
it should have good errors.  Cleanup is pending the SET/LOOKAHEAD and
SET-INFIX changes being integrated.