Variadic Quoting Infix Lambdas (a.k.a. "short function")
    >> printer: [a] -> print a + a * 2

    >> printer 10
    40

The body may be in a block, or if it's not a block then the code will
be read until a terminal is reached (e.g. ending parentheses, or end
of block, or an expression barrier).  Further elaboration at:

    https://trello.com/c/V3BVGAax

Technical restrictions prevent this from getting short to the point of
(a -> blah a blah), since infix functions intrinsically cannot quote
their left arguments.  However, with the <end> sensing, it is at least
possible to write (-> blah blah) and know there is nothing to the
left of the arrow, so the function can have no arguments.