Implement INFIX? as property of binding, not function
The rationale for this change and its generality is laid out on Trello:

https://trello.com/c/OvNE3GPM

Rather than have a separate OP! function type, this makes the choice
of a function to "look back" to find its first argument something that
is linked up at the time of setting the word.  Each SET or assignment
through a set word will reset this property.

There are some current technical limitations, but they should be
solvable--and this clears up a lot of user-facing concern, to the
extent that a user can be confident that if a function is assigned via
a SET that they performed (or is an argument to a function) that it
will not act as infix/postfix/etc.