writability check C++ only, kill VAL_INIT_WRITABLE_DEBUG
The cell formatting for writability check became important once some
cell values marking END were actually integers used for other purposes,
and not represenative of full cells.  However, stack-allocated C
variables carried uninitialized bits and needed to be manually cleared.

Since C++11 it has been legal to add constructors to structs and
keep their "plain old data" status (standard alignment, trivally movable
with memcpy, etc.):

http://stackoverflow.com/a/7189821/211160

Given the increasing reliance on the C++ build for debug checking in
general, it makes sense to use this ability to have REBVAL perform this
writability construction itself.  This won't cover REBVALs in series
or other places that are malloc'd, which still require a manual
call to INIT_CELL_WRITABLE_IF_DEBUG().

The result is a lifting of a burden at the callsites for this debug-only
purpose, with the consequence that C-based debug builds won't get the
check.  It also means that a C++-built debug binary can't have extension
code built separately pass in stack-allocated values from a C-built
binary (debug or release)...though there could be some kind of runtime
switch to turn it off in the library initialization.