Make APPLY and SPECIALIZE accept words/paths
The first implementation of NewApply made a kind-of silly choice to
quote its argument in order to get the name of any GET-WORDs, so as
to give more meaningful debug and backtrace information.  (As if you
just say APPLY :APPEND [...], APPLY receives an anonymous function and
does not know it is called append.)

This commit changes it to do the more logical thing, which is to add
a new feature that word-based APPLY callers can use...by naming the
function to call with a WORD! or PATH! if they have one.  The same
technique is included in SPECIALIZE.  This will not be retroactive to
old GET-WORD! applies, but a wrapper or variadic trick could be
translated into this trick for those interested.