Enhance legacy return-none-instead-of-void feature
This changes the rule for whether or not a none (BLANK!) is returned
from operations like SELECT and IF to be sensitive not just to the
legacy mode switch, but to the flag encoded on the array of running
code regarding where it came from.

Hence, turning on the legacy return mode will not affect the authoring
of mezzanine routines with different expectations.  The sensitivity to
whether the source was loaded before or after the r3-legacy mode was
turned on will determine the behavior.

Though not a perfect solution, it is an alternative (for the moment)
to rebinding the source to adapted constructs.  Long-term, rebinding is
the right way to handle the issue (though having a feature of this
kind could come in handy for things that cannot be solved that way.)