Make FIRST, SECOND, etc. use specialization of PICK
Specialization has firmed up to the point where it's a logical choice
for operations like the ordinal picks.  It has the advantage of only
making one function call, and generally speaking does not cost more
than the call itself.

Because specializations currently do not have their own specs, they do
not currently have their own descriptions.  For now, this slightly
improves that by making MOLD dump the exemplar FRAME! that is used to
pre-fill each call.  For this to get better HELP needs to be improved
to be aware of specialized refinements, and there needs to be an
option to "redescribe" the title of the function and its unspecialized
arguments.

This change was motivated by the observation that having these functions
implemented with their own bodies was leading the "legacy NONE" trigger
to not fire in response to the PICK, because it thought since they
were mezzanine they were written to new expectations.  While turning
them into specializations doesn't solve the general problem (there is
no general solution), it fixes this one...and gets the operations
using modern features.