not TRUE/NONE, not WORD/NONE but TRUE/FALSE refinements
This canonizes refinement values in both system-called functions and
frame-invoked functions to only allowing LOGIC! true and false.

This is a change from one of the earliest Ren-C adaptations, which was
to try and take advantage of making refinements their own WORD! in
order to ease chaining.  e.g. a function with a refinement /ONLY could
then chain that to APPEND by writing APPEND/:ONLY, where its none
status would not apply it and its word status would use it.

While that was a useful feature for its time, a lot has changed since.
With GROUP! executable in paths, one can write APPEND/(IF ONLY 'ONLY)
which is not much longer.  There are also other possibilities with
APPLY for doing proxying of words from the current frame conveniently.

What ultimately pushed the decision to TRUE and FALSE, was a technical
issue.  Consider the following:

    frame: make frame! :append
    frame/series: [1 2 3]
    frame/value: [4 5]
    frame/only: true

This is likely what a user would want to write, instead of having to
say `frame/only: 'only`.  If one is not making a copy of this frame
but just running the function with the values "as is", then some
point in the process would have to canonize the true to 'only.  That
canonization adds complexity as opposed to just checking for LOGIC!.

A similar line of argument can be applied to FALSE.  Ergonomically, it
would seem that one should be able to write FALSE, yet it would either
have to be canonized to a BLANK! or you make FALSE the only thing
accepted.

Trello notes here:

https://trello.com/c/H9I0261r