Clear frame flags when used by not-Do-Core
Frames are created sometimes by things like CASE that want to give
better debug information about where an error occurs (instead of
just pinning the error on CASE itself.)  Because these frames are
for message delivery and do not actually run the evaluator, the flags
were left unset.

However, valgrind noticed that a legacy switch which climbed an
arbitrary stack was checking the varargs flag on the frame.  This
was reading an uninitialized variable.  This sets it to 0 to be safe,
but the concept of these "error delivery frames" needs general review.
(R3-Alpha did not have such things, so they are new and haven't
gotten a thorough treatment).