Fix infix bug, reorganize code, enable "punctuation"
This is a reorganization trying to make sure all the bases are covered
in "lookahead" function handling.  As a major part of the change, it
involves recognizing that a REBFUN pointer to a function is not
sufficient to identify definitional returns--while the main loop
recognized this fact, the infix code didn't handle it.  This prevented
things like being able to create a specialization of RETURN that was
postfix--for instance.

So this cuts down on those repeated code paths, and adds the ability
for arity 0 functions used with lookahead to act as "punctuation"

    >> set/lookback 'stop does [print "stop" 1020]

    >> stop
    stop
    == 1020

    >> 1 2 3 stop
    stop
    == 1020

    >> 1 + stop
    ** Script error: stop punctuation can't serve as function argument

The commit also corrects an uninitialized boolean usage in the process.