Add PUNCTUATOR? and SET-PUNCTUATOR
This formalizes the naming of an arity 0 lookback function as a
"punctuator".  Also includes a tweak so that an infix function after
it cannot reach back and use its result, which makes it more
"punctuating" -- if its return were void then it would be very similar
to an expression barrier implemented in user code.