Soft quote directly into f->arg, simplify code
Soft quoting was optimized for the case of putting the quoted value into
f->out, which is only used if the evaluation throws.  Hence it had to
be copied into f->arg in the common case.

This reverses that bias and goes directly to f->arg, and simplifies the
code to make it easier to factor out for other soft-quoting cases.