Code reorganization and renamings (no effect)
This commit should not have any effect, and just does some light
reorganizations and renamings.

ET_NONE is changed to ET_INERT, to say that a value evaluates to itself.
The word "self" is already overloaded, and ET_LITERAL would be confusing
with LIT-WORD! etc... so "inert" seems about the best way of saying
that something has no evaulatior behavior.

The eval_type is saved in a variable instead of just being the condition
of a switch.  This is because it becomes useful to know what it is after
the evaluation (for infix quoting of left-hand arguments).

The param classes are renumbered so that a hard quote and a soft quote
share a common bit that can be tested for.

A section is added at the top of the evaluator for errors that are
specific to the evaluator, to start cutting down on the amount of
error-making code in Do_Core, but also to group the errors together
with cases where specific knowledge of a frame is needed.