New invariant for f->param (for lookback quoting)
When a set-word is evaluated, e.g. `x: 1 + 2`, it is intrinsically
necessary that the evaluator keep track of the SET-WORD! and its
binding, because it evaluates the expression and then sets it afterward.

This information could be made available to something like an infix
operator that quoted its left argument.  There has to be a stack
convention so that nested evaluations can look up one stack level
and see whether there's a pending set-word! or set-path! or not.

This commit adds the invariant.  It removes the free usage of f->param
for other things, which was previously a way for natives to do a GC
protect of a cell without a push/pop to the GUARD_VALUE stack.
Instead, it promises that f->param will be an END, a TYPESET, a
SET-WORD, or a SET-PATH.