Generalize "punctuates" concept to prefix functions
The feature by which 0-arity functions could stop themselves by being
consumed as the left hand argument of an infix function can only apply
to a function with no arguments (if they had arguments, it would be
their arguments deciding whether to be consumed, and then if those
arguments took arguments it would be their arguments...etc.)

However, the feature by which they could stop themselves as being
consumed as an argument from the left-hand side could technically be
applied to any ordinary prefix function with any number of arguments.
This feature is desirable for functions like PRINT for instance.
PRINT returns void, but as voids can be treated as "opting-out"...and
it's unlikely that the sequence:

    some-func a b c
    print "Hello"

...was intended to pass print in order to "opt out" of a 4th argument
to some-func.  More likely, it is a mistake.

This brings the left-side checking to any function by means of putting
a BAR! as the first element of the function spec, to turn it into a
function that "punctuates?"