Add in a `void` function
For aesthetic reasons, some people might prefer to write `return void`
instead of `return ()`.  It's not possible to fetch a void value from
a word, so that has to be a function...which makes it less efficient.
It also could be confusing, because unlike `get 'blank` or `get 'bar`
or `get 'true` the value that comes back would be a function.

But rather than be prescriptive that people have to say () in the source
if they don't like the way that looks, this standardizes the option
so everyone doesn't have to make this function themselves.