Add || userspace "expression barrier that runs one expr"
This adds an experimental form of expression barrier that is not built
in, but is a userspace function that leverages variadics and lookahead,
plus the notion of `<punctuates>`.  It tries to run only one expression
after it, e.g.

   >> 3 + 4 || 5 + 6
   11

   >> 3 + 4 || 5 + 6 7
   ** error

   >> 3 + 4 || 5 + 6 || 7
   7

While it works, it has an issue in terms of the definition of what it
should accept as counting as an "ok" terminal.  It clearly wants to
count itself and the `|>` and `<|` expressions, as well as BAR! and the
end of a block or group it's running.  But checking for literal words
and values isn't as good as using some property.

It uses PUNCTUATES? for now, but this exposes the limits of lookahead
on PATH!s, because if a path expression is seen in the lookahead it
isn't possible to tell if it's a punctuating function or not without
potentially running code.

It errs on the safe side and does not consider any paths to be valid
terminals, for now.