Don't pass voids via Do_Va C varlist calls
To make it easier to call the system support function for DO (that lets
it act upon FILE!, STRING!, TAG!, etc.) from C, it was at one time
changed to not use refinements but to use optional arguments.  This
is because identifying refinement arguments by position is brittle, and
using refinements is a little bit awkward in the C code when building
paths for function execution.

However, with voids no longer condoned allowed in blocks, the code that
"reified" C va_lists into series--in order to protect them from GC--
ran afoul of this rule.  For the moment, this changes the routine to
essentially take the refinements positionally.