More conservatively convert valist to array in GC
The test for whether or not to reify a valist was based on the index,
with the assumption that if the index was at the END_FLAG of input then
there would be no need to reify it.  This created a more complex
invariant, because it was not possible to assume in the GC that all
frames have source arrays.

For simplicity's sake, this goes ahead and reifies all valists--even
empty ones (which all reuse EMPTY_ARRAY anyway).