Move error case cleanup of frames to Fail_Core()
Following R3-Alpha's design, the cleanup code for when an error happened
was to keep tracking information in the "Rebol State", and use that
state as the only resource for cleaning up after a longjmp.

With call frames becoming more "nuanced", it was an unnecessary cost in
complexity--and a loss of verification--to be doing the cleanup after
the jump.  This does a limited amount of processing in response to the
Fail_Core call itself, when the call frames are still available to
examine.

Doing this opened up the opportunity to not manage a varlist of an
in-progress built frame, because it could be freed by the walk before
the error occurred.  It also allowed the chunk stack to become simpler
because there was no need to "slipstream" the context pointer into it.

The code is still a bit awkward in places, but is heavily asserted to
try and detect problems.