Get rid of ET_TERMINAL and improve arity 0 lookback
The ET_TERMINAL idea was broken, so this uses a DO_FLAG on the transient
lookahead_flags instead.  This also exposed a bug where the ANY_EVAL
logic had been thrown off by ET_INERT being bumped from the 0 slot.

Included in the fix is improved behavior and error messages for the
"terminal" behavior of arity 0 lookback functions.  They now are
willing to act as an "end" if they returned void, so that an infix
function which is end-tolerant can run after them in that case.

    >> set/lookback 'xxx func [left [<end> integer!] right [integer!]][
        print [:left "xxx" right]
    ]

    >> 10 xxx 20
    10 xxx 20

    >> xxx 20
    xxx 20

    >> set/lookback 'yyy does [1020]

    >> yyy xxx 20
    ** Script error: xxx isn't allowed 1st argument from previous call

    >> set/lookback 'zzz does []

    >> zzz xxx 20
    xxx 20