Mutation bugfix for common.r's to-c-name
The routine `to-c-name` was turning things like "+" into "plus" with
a switch statement.  However, it was not copying the strings.

As a consequence, callers who used `to-c-name` with the expectation that
they could modify the result (e.g. that it was acting like a FORM)
would wind up modifying the strings in the switch statement.

This is one good example of why there are likely strong good reasons for
considering source immutable by default.