Replace CALL_MODE_XXX with just eval type ET_XXX
With features emerging like being able to tell if a stack level above
the current one is currently fulfilling a SET-WORD! or SET-PATH!, the
code was becoming a bit twisty.  It was also difficult to think of
how many CALL_MODE_XXX there should be, and being somewhat "artificial"
there wasn't that much intuition to apply about those modes.

Since the ET_ eval type had to be chosen already, it provided a good
branch point for the frame's current stored operating mode.  From there,
the state of other frame fields could be used to determine whatever the
operating "nuances" of that state were.

Besides being clearer, this offers performance advantages, since the
eval_type calculation happening on every loop removes the need to worry
about resetting back to "mode 0".