QUOTE isn't `<opt>`, disables VALUE_FLAG_EVALUATED
It doesn't make sense to say that QUOTE (or any hard quoted function)
is able to take an optional parameter.  It can only receive void if it
is an `<end>`

This fixes a couple of other cases of optional hard quoted arguments,
and introduces an error at function creation time if more are made.