Overhaul infix dispatch to reuse main fulfill loop
This commit contains many improvements to infix handling.  It is based
on the realization that a generalized determination of the "first"
parameter to put an infix argument into needed to reuse the main loop
in order to get the proper behavior with things like specializations
and pure locals taken into account.

It also takes advantage of the new ability to know the `eval_type` of
a frame to allow quoting of left-hand arguments to infix functions.