[Linux] Rework copy-and-paste
So that it doesn't crash on big writing to clipboard. Crash happens
because the XChangeProperty returns an error when data is too big. When
data is big, it's better to use INCR mechanism doing it. However,
the possiblitity to have multiple content requestor makes it a little
bit too complicated (the content provider needs to maintain the state of
each content provider, and send data accordingly). So only copying using
INCR is implemented, and BadLength error is detected to avoid crashing.