Allow SPECIALIZE/APPLY of definitional returns
In order to specialize or apply functions, a frame must be created which
can uniquely identify the function to call.  Because all RETURNs and
LEAVEs use a "trick" so there's only one archetypal function for them,
they presented a problem.

This corrects the problem by on-demand creating a new paramlist, with
the "hacked up" REBVAL in its first slot.  Hence specialization of
these functions costs an extra 1 or 2-unit REBARR (cheap).