Simplify throwing to utilize the THROWN() bit
With definitional returns and exiting in general, it's necessary to
distinguish between setting a frame's "indexor" to THROWN_FLAG and
having a pending throw.  If the indexor is the only way of indicating
a throw happened, then there'd be no way to recover at an index if
a throw is caught.

Having tried different approaches and places to put the bit, having it
on the value itself as a THROWN bit is probably most convenient--
despite some feeling that the thrown bit was "wasting" a piece of
information that could be held as state in the evaluator.  It could be
held there...however, that would mean maintaining that state independent
of the value, which could become disconnected and would require setting
bits in two places instead of just having the throw state reset by
overwriting the value.

This commit includes several simplificaitons to try and make the code
readable, and also to prevent functions from catching exits during
parameter gathering (since the function has not completed argument
fulfillment, it should not be a candidate for exits).