Use END to initialize out slot, less GC protection
This changes the dynamic of GC protection of the f->out slot so that it
is only guarded during function execution.  The reason is not to save
time in the garbage collector (it's negligible), but rather to reduce
the number of cases where the slot has to contain valid initialized
data--as raw bits could crash the GC if it looked at them.

In addition to saving cycles in the evaluator, establishing a convention
that entry points to ET_FUNCTION should be responsible for initializing
f->out to END (or a lookback of a valid f->out, which will then be
turned into an end), there are several other advantages.  One advantage
is that an END marker will alert a /torture GC sweep if a movable
memory slot is used as the target of a DO, since it will see a stray
end marker.  Another is that since END is "out of band" for Rebol
evaluative results, natives can use this information to know if they've
ever written their output slot...which can save on needing to track
separate flags for that purpose (e.g. in WHILE?)